// src/core/types/pexels.ts
export interface I_Photo {
  id: number;
  width: number;
  height: number;
  url: string;
  photographer: string;
  photographer_url: string;
  photographer_id: number;
  avg_color: string;
  src: {
    original: string;
    large2x: string;
    large: string;
    medium: string;
    small: string;
    portrait: string;
    landscape: string;
    tiny: string;
  };
  liked: boolean;
  alt: string;
}

export interface I_PexelsResponse {
  page: number;
  per_page: number;
  photos: I_Photo[];
  total_results: number;
  next_page: string;
}